Move forms into dir components in keycloak-user-management
