+++ packages/keycloak-user-management/src/components/UserList/tests/utils.test.tsx
@@ -1,4 +1,4 @@
-import * as fixtures from '../../../forms/UserForm/tests/fixtures';
+import * as fixtures from '../../forms/UserForm/tests/fixtures';
 import { getDataFilters, getTableColumns } from '../utils';
 
 describe('components/UserList/utils/getDataFilters', () => {
